requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.2
torch==2.2.0
timm==0.9.0
torchvision==0.17.0
pillow==10.0.1
numpy==1.25.2
